Moves test into new e2e that has multiple backends
